+++ CHANGELOG.md
@@ -394,6 +394,7 @@ and this project adheres to
 - ✨(langfuse) allow user to score messages from LLM #6
 - ✨(onboarding) add activation code logic for launch #62
 - 💄(chat) add code highlighting for LLM responses #67
+- 🔧(evals) add run_evals management command
 
 [unreleased]: https://github.com/suitenumerique/conversations/compare/v0.0.19...main
 [0.0.19]: https://github.com/suitenumerique/conversations/releases/v0.0.19